requirements.txt
flask
requests
python-nmap
python-whois
reportlab